#505 move fiscal deal tree creation functions to fisc module
