+++ src/f00_instrument/file.py
@@ -347,3 +347,12 @@ def count_dirs_files(x_dir: str) -> int:
         num_files += len(files)  # Count files
 
     return num_dirs + num_files
+
+
+def get_level1_dirs(x_dir: str) -> list[str]:
+    """returns sorted list of all first level directorys"""
+    try:
+        level1_dirs = get_dir_file_strs(x_dir, include_dirs=True, include_files=False)
+        return sorted(list(level1_dirs.keys()))
+    except OSError as e:
+        return []
