+++ src/f11_world/world.py
@@ -50,7 +50,7 @@ from src.f10_etl.transformers import (
     etl_fisc_table2fisc_ote1_agg_csvs,
     etl_fisc_ote1_agg_csvs2jsons,
     etl_create_root_deal_nodes,
-    etl_create_deal_trees,
+    etl_create_fisc_deal_trees,
 )
 from dataclasses import dataclass
 from sqlite3 import connect as sqlite3_connect, Connection as sqlite3_Connection
@@ -223,8 +223,8 @@ class WorldUnit:
     def create_root_deal_nodes(self):
         etl_create_root_deal_nodes(self._fisc_mstr_dir)
 
-    def create_deal_trees(self):
-        etl_create_deal_trees(self._fisc_mstr_dir)
+    def create_fisc_deal_trees(self):
+        etl_create_fisc_deal_trees(self._fisc_mstr_dir)
 
     def mine_to_forecasts(self):  # sourcery skip: extract-method
         fisc_mstr_dir = create_path(self._world_dir, "fisc_mstr")
